refactor(ffmpeg): split the 1694-line plan_builder.go by pipeline stage
Pure same-package file reorganization (no call-site changes): the FFmpeg
plan_builder god-file is decomposed along the pipeline into cohesive siblings.

- plan_builder.go (1694 -> 88): orchestrator buildArgsWithPlan + plan types only
- plan_input.go:     input resolution + live-FPS resolution/probe
- plan_codec.go:     codec/HWAccel mapping + interlaced-path safety
- plan_serviceref.go: Enigma2 service-ref extraction + Safari runtime overrides
- plan_output.go:    live output/segment layout + HLS args
- encode_args.go:    VAAPI/NVENC/CPU/copy encoder args + rate-control + filters

No behavior change. go build + go vet clean; ffmpeg argv golden tests green;
golangci-lint --new-from-rev=origin/main = 0 new issues.
